+++ packages/stacked_cli/example/lib/app/app.bottomsheets.dart
@@ -1,0 +1,27 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+// **************************************************************************
+// StackedBottomsheetGenerator
+// **************************************************************************
+
+import 'package:stacked_services/stacked_services.dart';
+
+import 'app.locator.dart';
+import '../ui/bottom_sheets/notice/notice_sheet.dart';
+
+enum BottomSheetType {
+  notice,
+// @stacked-bottom-sheet-type
+}
+
+void setupBottomSheetUi() {
+  final bottomsheetService = locator<BottomSheetService>();
+
+  final Map<BottomSheetType, SheetBuilder> builders = {
+    BottomSheetType.notice: (context, request, completer) =>
+        NoticeSheet(request: request, completer: completer),
+    // @stacked-bottom-sheet-builder
+  };
+
+  bottomsheetService.setCustomSheetBuilders(builders);
+}
