+++ packages/stacked_cli/example/lib/app/app.dialogs.dart
@@ -1,0 +1,27 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+// **************************************************************************
+// StackedDialogGenerator
+// **************************************************************************
+
+import 'package:stacked_services/stacked_services.dart';
+
+import 'app.locator.dart';
+import '../ui/dialogs/info_alert/info_alert_dialog.dart';
+
+enum DialogType {
+  infoAlert,
+// @stacked-dialog-type
+}
+
+void setupDialogUi() {
+  final dialogService = locator<DialogService>();
+
+  final Map<DialogType, DialogBuilder> builders = {
+    DialogType.infoAlert: (context, request, completer) =>
+        InfoAlertDialog(request: request, completer: completer),
+    // @stacked-dialog-builder
+  };
+
+  dialogService.registerCustomDialogBuilders(builders);
+}
