+++ packages/stacked_cli/example/lib/app/app.router.dart
@@ -5,45 +5,45 @@
 // **************************************************************************
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'package:example/ui/views/home/home_view.dart' as _i3;
-import 'package:example/ui/views/startup/startup_view.dart' as _i2;
+import 'package:example/ui/views/home/home_view.dart' as _i2;
+import 'package:example/ui/views/startup/startup_view.dart' as _i3;
 import 'package:flutter/material.dart';
 import 'package:stacked/stacked.dart' as _i1;
 import 'package:stacked_services/stacked_services.dart' as _i4;
 
 class Routes {
-  static const startupView = '/startup-view';
-
   static const homeView = '/home-view';
 
+  static const startupView = '/startup-view';
+
   static const all = <String>{
-    startupView,
     homeView,
+    startupView,
   };
 }
 
 class StackedRouter extends _i1.RouterBase {
   final _routes = <_i1.RouteDef>[
     _i1.RouteDef(
-      Routes.startupView,
-      page: _i2.StartupView,
+      Routes.homeView,
+      page: _i2.HomeView,
     ),
     _i1.RouteDef(
-      Routes.homeView,
-      page: _i3.HomeView,
+      Routes.startupView,
+      page: _i3.StartupView,
     ),
   ];
 
   final _pagesMap = <Type, _i1.StackedRouteFactory>{
-    _i2.StartupView: (data) {
+    _i2.HomeView: (data) {
       return MaterialPageRoute<dynamic>(
-        builder: (context) => const _i2.StartupView(),
+        builder: (context) => const _i2.HomeView(),
         settings: data,
       );
     },
-    _i3.HomeView: (data) {
+    _i3.StartupView: (data) {
       return MaterialPageRoute<dynamic>(
-        builder: (context) => const _i3.HomeView(),
+        builder: (context) => const _i3.StartupView(),
         settings: data,
       );
     },
@@ -56,20 +56,6 @@ class StackedRouter extends _i1.RouterBase {
 }
 
 extension NavigatorStateExtension on _i4.NavigationService {
-  Future<dynamic> navigateToStartupView([
-    int? routerId,
-    bool preventDuplicates = true,
-    Map<String, String>? parameters,
-    Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)?
-        transition,
-  ]) async {
-    return navigateTo<dynamic>(Routes.startupView,
-        id: routerId,
-        preventDuplicates: preventDuplicates,
-        parameters: parameters,
-        transition: transition);
-  }
-
   Future<dynamic> navigateToHomeView([
     int? routerId,
     bool preventDuplicates = true,
@@ -84,14 +70,14 @@ extension NavigatorStateExtension on _i4.NavigationService {
         transition: transition);
   }
 
-  Future<dynamic> replaceWithStartupView([
+  Future<dynamic> navigateToStartupView([
     int? routerId,
     bool preventDuplicates = true,
     Map<String, String>? parameters,
     Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)?
         transition,
   ]) async {
-    return replaceWith<dynamic>(Routes.startupView,
+    return navigateTo<dynamic>(Routes.startupView,
         id: routerId,
         preventDuplicates: preventDuplicates,
         parameters: parameters,
@@ -106,6 +92,20 @@ extension NavigatorStateExtension on _i4.NavigationService {
         transition,
   ]) async {
     return replaceWith<dynamic>(Routes.homeView,
+        id: routerId,
+        preventDuplicates: preventDuplicates,
+        parameters: parameters,
+        transition: transition);
+  }
+
+  Future<dynamic> replaceWithStartupView([
+    int? routerId,
+    bool preventDuplicates = true,
+    Map<String, String>? parameters,
+    Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)?
+        transition,
+  ]) async {
+    return replaceWith<dynamic>(Routes.startupView,
         id: routerId,
         preventDuplicates: preventDuplicates,
         parameters: parameters,
