--- packages/stacked_cli/example/lib/enums/bottom_sheet_type.dart
@@ -1,4 +1,0 @@
-enum BottomSheetType {
-  notice,
-  // @stacked-bottom-sheet-type
-}
