--- packages/stacked_cli/example/lib/enums/dialog_type.dart
@@ -1,4 +1,0 @@
-enum DialogType {
-  infoAlert,
-  // @stacked-dialog-type
-}
