+++ packages/stacked_cli/example/test/viewmodels/notice_sheet_model_test.dart
@@ -1,6 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:example/app/app.locator.dart';
-import 'package:example/ui/bottom_sheets/notice/notice_sheet_model.dart';
 
 import '../helpers/test_helpers.dart';
